+++ Paramount.Betterclassifieds.DataService/AdoModels/BetterclassifiedDataService.cs
@@ -9,17 +9,13 @@ namespace Paramount.Betterclassifieds.DataService
     [Obsolete]
     public class BetterclassifiedDataService
     {
-        private const string ConfigSection = @"paramount/services";
-        private const string ConfigKey = "Betterclassifieds";
-        
         public static List<ExpiredAdRow> GetExpiredAdByLastEdition(DateTime editionDate)
         {
-            var df = new DatabaseProxy("psp_Betterclassified_GetLineAdBookingByLastEdition", ConfigSection, ConfigKey);
+            var df = new DatabaseProxy("psp_Betterclassified_GetLineAdBookingByLastEdition", "ClassifiedConnection");
             df.AddParameter("@EditionDate", editionDate);
             var dt = df.ExecuteQuery().Tables[0];
             var list = (from DataRow item in dt.Rows select new ExpiredAdRow(item)).ToList();
             return list;
         }
-
     }
 }