+++ WebApplication/iFlog/payment/notify.aspx.vb
@@ -13,7 +13,7 @@ Partial Public Class notify
 
     Public Shared ReadOnly Property Settings() As PayPalSettings
         Get
-            Return ConfigurationManager.GetSection("paypal")
+            Return ConfigurationManager.GetSection("paypal-classic")
         End Get
     End Property