+++ WebApplication/iFlog/payment/pp.aspx.vb
@@ -7,7 +7,7 @@ Partial Public Class pp
     Inherits System.Web.UI.Page
     Public Shared ReadOnly Property Settings() As PayPalSettings
         Get
-            Return ConfigurationManager.GetSection("paypal")
+            Return ConfigurationManager.GetSection("paypal-classic")
         End Get
     End Property